Move auxengine code to auxengine.cc
It's still part of the Search/SearchWorker classes.
